+++ src/test/java/org/plumelib/reflection/TestReflectionPlume.java
@@ -53,6 +53,16 @@ public final class TestReflectionPlume {
   }
 
   @Test
+  public void test_nameWithoutPackage() {
+    assertEquals("String", ReflectionPlume.nameWithoutPackage(String.class));
+    assertEquals("Map.Entry", ReflectionPlume.nameWithoutPackage(java.util.Map.Entry.class));
+    assertEquals("TestReflectionPlume.Inner", ReflectionPlume.nameWithoutPackage(Inner.class));
+    assertEquals(
+        "TestReflectionPlume.Inner.InnerInner",
+        ReflectionPlume.nameWithoutPackage(Inner.InnerInner.class));
+  }
+
+  @Test
   public void test_isSubtype() {
     // private boolean isSubtype(Class<?> sub, Class<?> sup) {
     assertTrue(ReflectionPlume.isSubtype(Integer.class, Integer.class));