+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.2
 
+- New methods in class ReflectionPlume:
+   - nameWithoutPackage
 - New methods in class Signatures:
    - isArrayWithoutPackage
    - isBinaryNameWithoutPackage